refactor: extract traffic fine logic
