+++ supabase/lib/supabaseClient.ts
@@ -1,0 +1,7 @@
+import { createClient } from 'https://esm.sh/@supabase/supabase-js@2.7.1';
+
+export function getSupabaseClient() {
+  const supabaseUrl = Deno.env.get('SUPABASE_URL') || '';
+  const supabaseKey = Deno.env.get('SUPABASE_SERVICE_ROLE_KEY') || '';
+  return createClient(supabaseUrl, supabaseKey);
+}
