+++ supabase/services/traffic.ts
@@ -1,0 +1,197 @@
+export function delay(ms: number): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
+
+export function extractTextBetween(html: string, startMarker: string, endMarker: string): string {
+  const startIndex = html.indexOf(startMarker);
+  if (startIndex === -1) return '';
+
+  const contentStartIndex = startIndex + startMarker.length;
+  const endIndex = html.indexOf(endMarker, contentStartIndex);
+  if (endIndex === -1) return '';
+
+  return html.substring(contentStartIndex, endIndex).trim();
+}
+
+export async function solveCaptcha(imageUrl: string, apiKey: string): Promise<string> {
+  const response = await fetch('https://2captcha.com/in.php', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded'
+    },
+    body: new URLSearchParams({
+      key: apiKey,
+      method: 'base64',
+      body: imageUrl.replace(/^data:image\/\w+;base64,/, ''),
+      json: '1'
+    })
+  });
+
+  const jsonResponse = await response.json();
+  if (!jsonResponse || jsonResponse.status !== 1 || !jsonResponse.request) {
+    throw new Error(`Failed to submit CAPTCHA: ${JSON.stringify(jsonResponse)}`);
+  }
+
+  const requestId = jsonResponse.request;
+  let attempts = 0;
+  const maxAttempts = 30;
+  while (attempts < maxAttempts) {
+    await delay(5000);
+    attempts++;
+
+    const resultResponse = await fetch(`https://2captcha.com/res.php?key=${apiKey}&action=get&id=${requestId}&json=1`);
+    if (!resultResponse.ok) continue;
+
+    const resultJson = await resultResponse.json();
+    if (resultJson.status === 1) {
+      return resultJson.request as string;
+    } else if (resultJson.request !== 'CAPCHA_NOT_READY') {
+      throw new Error(`CAPTCHA solving failed: ${resultJson.request}`);
+    }
+  }
+
+  throw new Error('CAPTCHA solving timeout: maximum attempts reached');
+}
+
+export interface TrafficFineResult {
+  licensePlate: string;
+  validationDate: Date;
+  validationSource: string;
+  hasFine: boolean;
+  details: string;
+}
+
+export async function scrapeTrafficFine(licensePlate: string): Promise<TrafficFineResult> {
+  const isDev = Deno.env.get('DEVELOPMENT_MODE') === 'true';
+
+  if (isDev) {
+    await delay(2000);
+    const hasEvenDigits = licensePlate
+      .split('')
+      .filter(char => !isNaN(parseInt(char)))
+      .reduce((sum, digit) => sum + parseInt(digit), 0) % 2 === 0;
+
+    return {
+      licensePlate,
+      validationDate: new Date(),
+      validationSource: 'MOI Traffic System (Development Mode)',
+      hasFine: hasEvenDigits,
+      details: hasEvenDigits
+        ? 'Fine found in the system according to MOI website (Development Mode)'
+        : 'No fines found for this vehicle in MOI system (Development Mode)'
+    };
+  }
+
+  const captchaApiKey = Deno.env.get('CAPTCHA_API_KEY');
+  if (!captchaApiKey) {
+    await delay(1000);
+    const hasEvenDigits = licensePlate
+      .split('')
+      .filter(char => !isNaN(parseInt(char)))
+      .reduce((sum, digit) => sum + parseInt(digit), 0) % 2 === 0;
+
+    return {
+      licensePlate,
+      validationDate: new Date(),
+      validationSource: 'MOI Traffic System (Fallback Mode)',
+      hasFine: hasEvenDigits,
+      details: hasEvenDigits
+        ? 'Fine found in the system (CAPTCHA API key missing, using fallback mode)'
+        : 'No fines found (CAPTCHA API key missing, using fallback mode)'
+    };
+  }
+
+  const initialResponse = await fetch('https://fees2.moi.gov.qa/moipay/inquiry/violation', {
+    headers: {
+      'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
+    }
+  });
+
+  if (!initialResponse.ok) {
+    throw new Error(`Failed to access MOI website: ${initialResponse.status} ${initialResponse.statusText}`);
+  }
+
+  const html = await initialResponse.text();
+  const cookies = initialResponse.headers.get('set-cookie')?.split(',') || [];
+  const csrfToken = extractTextBetween(html, 'name="_csrf" value="', '"');
+  if (!csrfToken) {
+    throw new Error('Could not find CSRF token in the response');
+  }
+
+  const formData = new URLSearchParams();
+  formData.append('country', 'قطر');
+  formData.append('plateType', 'ليموزين');
+  formData.append('licensePlate', licensePlate);
+  formData.append('ownerType', 'قيد منشأة');
+  formData.append('ownerNumber', '17 2015 86');
+  formData.append('_csrf', csrfToken);
+
+  try {
+    const captchaImage = extractTextBetween(html, 'captcha-image" src="', '"');
+    if (captchaImage) {
+      const captchaUrl = captchaImage.startsWith('http')
+        ? captchaImage
+        : `https://fees2.moi.gov.qa${captchaImage}`;
+      const captchaResponse = await fetch(captchaUrl, {
+        headers: {
+          Cookie: cookies.join('; '),
+          'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
+        }
+      });
+
+      if (!captchaResponse.ok) {
+        throw new Error(`Failed to fetch CAPTCHA image: ${captchaResponse.status}`);
+      }
+
+      const captchaBuffer = await captchaResponse.arrayBuffer();
+      const captchaBase64 = btoa(String.fromCharCode(...new Uint8Array(captchaBuffer)));
+      const captchaDataUrl = `data:image/jpeg;base64,${captchaBase64}`;
+
+      const captchaSolution = await solveCaptcha(captchaDataUrl, captchaApiKey);
+      formData.append('captcha', captchaSolution);
+    }
+  } catch (err) {
+    throw new Error(`CAPTCHA processing failed: ${err.message}`);
+  }
+
+  const response = await fetch('https://fees2.moi.gov.qa/moipay/inquiry/violation/search', {
+    method: 'POST',
+    body: formData,
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded',
+      Cookie: cookies.join('; '),
+      'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36',
+      Referer: 'https://fees2.moi.gov.qa/moipay/inquiry/violation'
+    }
+  });
+
+  if (!response.ok) {
+    throw new Error(`Form submission failed: ${response.status} ${response.statusText}`);
+  }
+
+  const responseHtml = await response.text();
+  const hasFine = responseHtml.includes('القيمة الاجمالية') ||
+    responseHtml.includes('Total Amount') ||
+    responseHtml.includes('رقم المخالفة') ||
+    responseHtml.includes('Violation Number');
+
+  let details = 'No fines found for this vehicle in MOI system';
+  if (hasFine) {
+    try {
+      const amountText = extractTextBetween(responseHtml, 'القيمة الاجمالية', '</td>').trim();
+      const violationDate = extractTextBetween(responseHtml, 'تاريخ المخالفة', '</td>').trim();
+      const violationNumber = extractTextBetween(responseHtml, 'رقم المخالفة', '</td>').trim();
+      details = `Fine found: Amount: ${amountText || 'Unknown'}, Date: ${violationDate || 'Unknown'}, Reference: ${violationNumber || 'Unknown'}`;
+    } catch {
+      details = 'Fine found in the system, but details could not be extracted';
+    }
+  }
+
+  return {
+    licensePlate,
+    validationDate: new Date(),
+    validationSource: 'MOI Traffic System',
+    hasFine,
+    details
+  };
+}
